+++ src/components/TextInputFormix.tsx
@@ -4,14 +4,14 @@ import { Entypo as Icon } from '@expo/vector-icons';
 
 
 export default function TextInput({ icon, error, touched, ...otherProps }) {
-  const validationColor = !touched ? '#223e4b' : error ? '#FF5A5F' : '#223e4b';
+  const validationColor = !touched ? '#ff4b6e' : error ? '#FF0000' : '#ff4b6e';
   return (
     <View
       style={{
         flexDirection: 'row',
         alignItems: 'center',
         height: 48,
-        borderRadius: 8,
+        borderRadius: 25,
         borderColor: validationColor,
         borderWidth: StyleSheet.hairlineWidth,
         padding: 8,
